refactor(webapp): decompose git-commands.ts by subcommand (#1311)
Split the 2,967-line GitCommands class into per-subcommand modules under
git/commands/, each receiving a GitCommandContext instead of `this`.
git-commands.ts now holds only the constructor, shared per-invocation state,
auth/token/author helpers, and the execute() dispatch (484 lines). Behavior is
unchanged.
